:zap: experimental getters
