+++ src/environment.ts
@@ -10,7 +10,9 @@ const NODE_URL = 'wss://statemine-rpc.polkadot.io';
 const DEV_ARCHIVE_URL = lookupArchive('statemint', {release: 'FireSquid'} );
 const DEV_NODE_URL = 'wss://statemint-rpc.polkadot.io';
 
-const isProd = CHAIN === 'statemine';
+export const isProd = CHAIN === 'statemine';
+
+console.log(`Using ${CHAIN} chain ${isProd ? 'production' : 'development'} environment`);
 
 export const getArchiveUrl = (): string => (isProd ? ARCHIVE_URL : DEV_ARCHIVE_URL);
 export const getNodeUrl = (): string => (isProd ? NODE_URL : DEV_NODE_URL);
